third_party/WebKit/LayoutTests/fast/dom/minor-dom-gc.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
function createTree() {
    var div = document.createElement("div");
    div.appendChild(document.createElement("div"));
    div.appendChild(document.createElement("div"));
    div.appendChild(document.createElement("div"));
}
gc();
var num1 = internals.numberOfLiveNodes();
createTree();
var num2 = internals.numberOfLiveNodes();
minorGC();
var num3 = internals.numberOfLiveNodes();
shouldBe('num2', 'num1 + 4');
shouldBe('num3', 'num1');
</script>
</body>
</html>
